+++ lib/arachni/element/capabilities/with_node.rb
@@ -6,44 +6,19 @@
     web site for more information on licensing and terms of use.
 =end
 
+require_relative 'with_source'
+
 module Arachni
 module Element::Capabilities
 
 # @author Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>
 module WithNode
-
-    # @return   [String]
-    #   HTML code for the element.
-    attr_accessor :html
-
-    def initialize( options )
-        super
-        self.html = options[:html].freeze
-    end
-
-    def html=( s )
-        @html = (s ? s.recode.freeze : s)
-    end
+    include WithSource
 
     # @return [Nokogiri::XML::Element]
     def node
-        return if !@html
-        Nokogiri::HTML.fragment( @html.dup ).children.first
-    end
-
-    def to_h
-        super.merge( html: html )
-    end
-
-    def dup
-        copy_with_node( super )
-    end
-
-    private
-
-    def copy_with_node( other )
-        other.html = html
-        other
+        return if !@source
+        Nokogiri::HTML.fragment( @source.dup ).children.first
     end
 
 end