+++ lib/arachni/issue.rb
@@ -297,7 +297,7 @@ class Issue
             h.delete( :variation )
         else
             if variation?
-                h[:vector].delete :html
+                h[:vector].delete :source
                 h[:vector].delete :type
                 h[:vector].delete :url
                 h[:vector].delete :action