+++ spec/support/factories/element/xml.rb
@@ -1,0 +1,5 @@
+Factory.define :xml, class: Arachni::Element::XML,
+               options: {
+                   url:    'http://test.com',
+                   source: '<input>value</input>'
+               }